Add open browser on Windows for preview
Why was this change made?

The browser does not open when executing the `-o` flag when running the
`preview` command on Windows.

What functionality was changed?

In order to support this need, the following changes have been made:
- Add a file called `command.go`.
- Modify `preview.go` to use the openURL command in the `command.go` file.

What are the side effects of the change outside of this repository?

This has not yet been validated on Mac OS or Linux distributions.
Additional testing should be performed on these platforms before merging
into master.