Add CLI flags to output well-formatted optional question attributes.
This change is backward compatible with the current learn interface.

This commit adds three new flags to the `learn md` command:

- `--with-explanation/-e` (bool)  which creates explanation attributes
  for correct and incorrect responses
- `--with-rubric/-r` (bool) which creates rubric attributes
- `--with-hints/-n` (int) which creates the indicated number of hint
  attributes

Without the minimal flag, the comments are preserved in the question
as it previously worked. If a user specifies one of the above flags,
then the corresponding comment is removed from the final output.